\name{rstata-package}
\alias{rstata-package}
\alias{rstata}
\docType{package}
\title{
\packageTitle{rstata}
}
\description{
\packageDescription{rstata}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rstata}
\packageIndices{rstata}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rstata}

Maintainer: \packageMaintainer{rstata}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
