% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstata.R
\docType{package}
\name{rstata}
\alias{rstata}
\alias{rstata-package}
\title{rstata: An implementation of Stata's ado language in R.}
\description{
The rstata package provides an interpreter for Stata's ado language built on
top of R. Loops, macros, data manipulation commands and statistics commands
are all supported, as are multiple ways to embed R code and use R for writing
ado- language commands. (The language this package implements has about the
same relationship to Stata that various dialects of SQL have to each other.)
This package is not in any way affiliated with or endorsed by StataCorp.
}
\section{Functionality overview}{

TODO
}

